DeepSeek-OCR-2/requirements_api.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.46.3
tokenizers==0.20.3
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
PyMuPDF
img2pdf
einops
easydict
addict
Pillow
numpy

